+++ src/TestUtils/TestDatasystemProviderService.php
@@ -11,57 +11,64 @@ use Symfony\Component\HttpFoundation\Response;
 
 class TestDatasystemProviderService implements DatasystemProviderServiceInterface
 {
-    public static $data = [];
+    private array $data;
 
-    public static function hasFile(string $internalBucketId, string $fileId): bool
+    public function __construct()
     {
-        return isset(self::$data[$internalBucketId][$fileId]);
+        $this->data = [];
     }
 
-    public static function isContentEqual(string $internalBucketId, string $fileId, File $fileToCompare): bool
+    public function hasFile(string $internalBucketId, string $fileId): bool
     {
-        $file = self::$data[$internalBucketId][$fileId] ?? null;
-
-        return $file !== null && $file->getContent() === $fileToCompare->getContent();
+        return isset($this->data[$internalBucketId][$fileId]);
     }
 
     public function saveFile(string $internalBucketId, string $fileId, File $file): void
     {
-        self::$data[$internalBucketId][$fileId] = $file;
+        $this->data[$internalBucketId][$fileId] = $file;
     }
 
     public function getBinaryResponse(string $internalBucketId, string $fileId): Response
     {
-        if (!isset(self::$data[$internalBucketId][$fileId])) {
+        if (!isset($this->data[$internalBucketId][$fileId])) {
             throw new \RuntimeException();
         }
 
         // build binary response
-        $response = new BinaryFileResponse(self::$data[$internalBucketId][$fileId]->getRealPath());
+        $response = new BinaryFileResponse($this->data[$internalBucketId][$fileId]->getRealPath());
 
         return $response;
     }
 
     public function removeFile(string $internalBucketId, string $fileId): void
     {
-        unset(self::$data[$internalBucketId][$fileId]);
-    }
-
-    public function getSumOfFilesizesOfBucket(string $internalBucketId): int
-    {
-        $sumOfFileSizes = 0;
-        $files = self::$data[$internalBucketId] ?? [];
-        foreach ($files as $file) {
-            $sumOfFileSizes += $file->getFileSize();
+        if (!isset($this->data[$internalBucketId][$fileId])) {
+            throw new \RuntimeException();
         }
 
-        return $sumOfFileSizes;
+        unset($this->data[$internalBucketId][$fileId]);
     }
 
-    public function getNumberOfFilesInBucket(string $internalBucketId): int
+    public function listFiles(string $internalBucketId): iterable
     {
-        $files = self::$data[$internalBucketId] ?? [];
+        return array_keys($this->data[$internalBucketId] ?? []);
+    }
 
-        return count($files);
+    public function getFileSize(string $internalBucketId, string $fileId): int
+    {
+        if (!isset($this->data[$internalBucketId][$fileId])) {
+            throw new \RuntimeException();
+        }
+
+        return $this->data[$internalBucketId][$fileId]->fileSize();
+    }
+
+    public function getFileHash(string $internalBucketId, string $fileId): string
+    {
+        if (!isset($this->data[$internalBucketId][$fileId])) {
+            throw new \RuntimeException();
+        }
+
+        return hash('sha256', $this->data[$internalBucketId][$fileId]->getContent());
     }
 }