+++ tests/CurlGetTest.php
@@ -12,7 +12,6 @@ use Dbp\Relay\BlobBundle\Configuration\ConfigurationService;
 use Dbp\Relay\BlobBundle\Helper\SignatureUtils;
 use Dbp\Relay\BlobBundle\Service\BlobService;
 use Dbp\Relay\BlobBundle\TestUtils\BlobApiTest;
-use Dbp\Relay\BlobBundle\TestUtils\TestDatasystemProviderService;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Dbp\Relay\CoreBundle\TestUtils\TestClient;
 use Dbp\Relay\CoreBundle\TestUtils\UserAuthTrait;
@@ -20,7 +19,6 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Uid\Uuid;
 
 /**
@@ -97,7 +95,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('GET', $url, $options);
 
             $this->assertEquals(200, $response->getStatusCode());
@@ -203,7 +200,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var \ApiPlatform\Symfony\Bundle\Test\Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
             $this->assertEquals(200, $response->getStatusCode());
 
@@ -285,7 +281,6 @@ class CurlGetTest extends ApiTestCase
 
             $token = SignatureUtils::createSignature($secret, $payload);
 
-            /** @var \ApiPlatform\Symfony\Bundle\Test\Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
             $this->assertEquals(200, $response->getStatusCode());
 
@@ -326,7 +321,6 @@ class CurlGetTest extends ApiTestCase
 
                 $token = SignatureUtils::createSignature($secret, $payload);
 
-                /** @var Response $response */
                 $response = $client->request('DELETE', $url.'&sig='.$token, $options);
 
                 $this->assertEquals(204, $response->getStatusCode());
@@ -353,7 +347,6 @@ class CurlGetTest extends ApiTestCase
 
             $client = $this->setUpTestClient();
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
 
             $this->assertEquals(200, $response->getStatusCode());
@@ -443,8 +436,9 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            $this->assertArrayHasKey($this->files[0]['uuid'], TestDatasystemProviderService::$data[$bucket->getInternalBucketId()], 'File data not in dummy store.');
-            /** @var Response $response */
+            $provider = $blobService->getDatasystemProvider($fileData);
+            $this->assertTrue($provider->hasFile($fileData->getInternalBucketId(), $this->files[0]['uuid']));
+
             $response = $client->request('GET', $url, $options);
             $this->assertEquals(200, $response->getStatusCode());
             // TODO: further checks...
@@ -467,7 +461,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('DELETE', $url.'&sig='.$token, $options);
             $this->assertEquals(204, $response->getStatusCode());
             // TODO: further checks...
@@ -491,7 +484,6 @@ class CurlGetTest extends ApiTestCase
             $token = SignatureUtils::createSignature($secret, $payload);
             $client = $this->setUpTestClient();
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
 
             $this->assertEquals(200, $response->getStatusCode());
@@ -537,7 +529,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
 
             $this->assertEquals(403, $response->getStatusCode());
@@ -574,7 +565,6 @@ class CurlGetTest extends ApiTestCase
 
             $token = SignatureUtils::createSignature($secret, $payload);
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, []);
 
             $this->assertEquals(404, $response->getStatusCode());
@@ -596,7 +586,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('DELETE', $url.'&sig='.$token, $options);
 
             $this->assertEquals(404, $response->getStatusCode());
@@ -701,7 +690,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
 
             $this->assertEquals(405, $response->getStatusCode());
@@ -802,7 +790,6 @@ class CurlGetTest extends ApiTestCase
                     ],
                 ];
 
-                /** @var Response $response */
                 $response = $client->request('GET', $url.'&sig='.$token, $options);
 
                 $this->assertEquals(405, $response->getStatusCode());
@@ -1025,7 +1012,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var \ApiPlatform\Symfony\Bundle\Test\Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options2);
             $this->assertEquals(200, $response->getStatusCode());
             // check if fileName was indeed changed
@@ -1218,7 +1204,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
             $this->assertEquals(200, $response->getStatusCode());
             // check if the one created element is there
@@ -1330,10 +1315,6 @@ class CurlGetTest extends ApiTestCase
                         'ucs' => $this->generateSha256ChecksumFromUrl($baseUrl),
                     ];
 
-                    if ($i !== count($params) - 1) {
-                        $baseUrl = $baseUrl.$connector.$params[$j];
-                    }
-
                     $token = SignatureUtils::createSignature($secret, $payload);
 
                     $options = [
@@ -1345,7 +1326,6 @@ class CurlGetTest extends ApiTestCase
 
                     $client = $this->setUpTestClient();
 
-                    /** @var Response $response */
                     $response = $client->request($action, $baseUrl.$token, $options);
                     $this->assertEquals(400, $response->getStatusCode());
                 }
@@ -1387,10 +1367,6 @@ class CurlGetTest extends ApiTestCase
                         'ucs' => $this->generateSha256ChecksumFromUrl($baseUrl),
                     ];
 
-                    if ($i !== count($params) - 1) {
-                        $baseUrl = $baseUrl.$connector.$params[$j];
-                    }
-
                     $token = SignatureUtils::createSignature($secret, $payload);
 
                     $file = new UploadedFile($this->files[0]['path'], $this->files[0]['name']);
@@ -1409,7 +1385,6 @@ class CurlGetTest extends ApiTestCase
 
                     $client = $this->setUpTestClient();
 
-                    /** @var Response $response */
                     $response = $client->request($action, $baseUrl.$token, $options);
                     $this->assertEquals(400, $response->getStatusCode());
                 }
@@ -1449,10 +1424,6 @@ class CurlGetTest extends ApiTestCase
                     'ucs' => $this->generateSha256ChecksumFromUrl($baseUrl),
                 ];
 
-                if ($i !== count($params) - 1) {
-                    $baseUrl = $baseUrl.$connector.$params[$j];
-                }
-
                 $token = SignatureUtils::createSignature($secret, $payload);
 
                 $options = [
@@ -1488,7 +1459,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request('POST', $baseUrl.$token, $options);
                 $this->assertEquals(400, $response->getStatusCode());
             }
@@ -1523,10 +1493,6 @@ class CurlGetTest extends ApiTestCase
                     $connector = '&';
                 }
 
-                if ($i !== count($params) - 1) {
-                    $baseUrl = $baseUrl.$connector.$params[$j];
-                }
-
                 $options = [
                     'headers' => [
                         'Authorization' => 'Bearer 42',
@@ -1537,7 +1503,6 @@ class CurlGetTest extends ApiTestCase
                 ];
 
                 $client = $this->setUpTestClient();
-                /** @var Response $response */
                 $response = $client->request('PATCH', $baseUrl, $options);
                 $this->assertEquals(400, $response->getStatusCode());
             }
@@ -1633,7 +1598,6 @@ class CurlGetTest extends ApiTestCase
                 ];
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(403, $response->getStatusCode());
             }
@@ -1653,7 +1617,6 @@ class CurlGetTest extends ApiTestCase
                 $token = SignatureUtils::createSignature($secret, $payload);
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(403, $response->getStatusCode());
             }
@@ -1677,7 +1640,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request('POST', $baseUrl.'&sig='.$token,
                     [
                         'headers' => [
@@ -1796,7 +1758,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(403, $response->getStatusCode());
             }
@@ -1820,7 +1781,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(403, $response->getStatusCode());
             }
@@ -1847,7 +1807,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request('POST', $baseUrl.'&sig='.$token,
                     [
                         'headers' => [
@@ -1964,7 +1923,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(403, $response->getStatusCode());
             }
@@ -1985,7 +1943,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(403, $response->getStatusCode());
             }
@@ -2009,7 +1966,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request('POST', $baseUrl.'&sig='.$token,
                     [
                         'headers' => [
@@ -2126,7 +2082,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(400, $response->getStatusCode());
             }
@@ -2147,7 +2102,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request($action, $baseUrl.'&sig='.$token, $options);
                 $this->assertEquals(400, $response->getStatusCode());
             }
@@ -2171,7 +2125,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request('POST', $baseUrl.'&sig='.$token,
                     [
                         'headers' => [
@@ -2272,7 +2225,6 @@ class CurlGetTest extends ApiTestCase
                 ],
             ];
 
-            /** @var Response $response */
             $response = $client->request('GET', $url.'&sig='.$token, $options);
             $this->assertEquals(200, $response->getStatusCode());
             $this->assertEquals('string', gettype($response->getContent()));
@@ -2377,10 +2329,6 @@ class CurlGetTest extends ApiTestCase
                     'ucs' => $this->generateSha256ChecksumFromUrl($baseUrl),
                 ];
 
-                if ($i !== count($params) - 1) {
-                    $baseUrl = $baseUrl.$connector.$params[$j];
-                }
-
                 $token = SignatureUtils::createSignature($secret, $payload);
 
                 $file = new UploadedFile($this->files[0]['path'], $this->files[0]['name']);
@@ -2399,7 +2347,6 @@ class CurlGetTest extends ApiTestCase
 
                 $client = $this->setUpTestClient();
 
-                /** @var Response $response */
                 $response = $client->request('GET', $baseUrl.$token, $options);
                 $this->assertEquals(400, $response->getStatusCode());
             }
@@ -2426,7 +2373,6 @@ class CurlGetTest extends ApiTestCase
 
             $client = $this->setUpTestClient();
 
-            /** @var Response $response */
             $response = $client->request('GET', $baseUrl.'&sig='.$token, $options);
             $this->assertEquals(403, $response->getStatusCode());